Сделал отдельный модуль инициализации страницы
