+++ js/main.js
@@ -1,8 +1,8 @@
 import { adFormHandler } from './form.js';
 import { mapFormHandler } from './filter.js';
-import { initMap } from './map.js';
+import { initPage } from './page.js';
 
 adFormHandler(document.querySelector('.ad-form'));
 mapFormHandler(document.querySelector('.map__filters'));
 
-initMap();
+initPage();
